requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1

# ML Libraries
lightgbm==4.3.0
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.0.2

# Monitoring & Logging
boto3==1.34.0
watchtower==3.0.1
prometheus-client==0.19.0

# Data Storage
redis==6.1.1
pyathena==3.0.0
pyarrow==14.0.1

# Utilities
httpx==0.27.0
python-multipart==0.0.9
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.27.0

